优化Jenkins的Docker构建脚本init-jenkins.sh

+++ src/jenkins/docker/init-jenkins.sh
@@ -1,21 +1,25 @@
 #!/bin/bash
 
-/sbin/tini -- /usr/local/bin/jenkins.sh "$@" &
+if ! [ -e "${JENKINS_HOME}/secrets/initialAdminPassword" ]; then
 
-while ! [ -e "${JENKINS_HOME}/hudson.model.UpdateCenter.xml" ]; do
-  sleep 1
-done
+  /sbin/tini -- /usr/local/bin/jenkins.sh "$@" &
 
-echo "开始替换国内源"
+  while ! [ -e "${JENKINS_HOME}/hudson.model.UpdateCenter.xml" ]; do
+    sleep 1
+  done
 
-# shellcheck disable=SC2009
-JENKINS_PROCESS_ID=$(ps -ef | grep jenkins.war | grep -v grep | awk '{print $2}')
+  echo "开始替换国内源"
 
-echo "进程id为${JENKINS_PROCESS_ID}"
+  # shellcheck disable=SC2009
+  JENKINS_PROCESS_ID=$(ps -ef | grep jenkins.war | grep -v grep | awk '{print $2}')
 
-#echo "${JENKINS_HOME}"
-sed -i 's/https:\/\/updates.jenkins.io/https:\/\/mirrors.tuna.tsinghua.edu.cn\/jenkins\/updates/g' "${JENKINS_HOME}/hudson.model.UpdateCenter.xml"
+  echo "进程id为${JENKINS_PROCESS_ID}"
 
-kill -9 "${JENKINS_PROCESS_ID}"
+  #echo "${JENKINS_HOME}"
+  sed -i 's/https:\/\/updates.jenkins.io/https:\/\/mirrors.tuna.tsinghua.edu.cn\/jenkins\/updates/g' "${JENKINS_HOME}/hudson.model.UpdateCenter.xml"
+
+  kill -9 "${JENKINS_PROCESS_ID}"
+
+fi
 
 /sbin/tini -- /usr/local/bin/jenkins.sh "$@"
